refactor(text-input): move folder, update type file dir
