+++ src/data-display/popover/PPopover.stories.mdx
@@ -4,7 +4,7 @@ import { reactive, toRefs } from 'vue';
 import { getPopoverArgTypes } from './story-helper';
 import PPopover from '@/data-display/popover/PPopover';
 import PButton from '../../inputs/buttons/button/PButton';
-import { faker } from '@faker-js/faker'; import PTextInput from '../../inputs/input/PTextInput';
+import { faker } from '@faker-js/faker'; import PTextInput from '../../inputs/input/text-input/PTextInput.vue';
 
 <Meta title='Data Display/Popover' parameters={{
     design: {
