+++ src/data-display/tree/PTree.stories.mdx
@@ -3,7 +3,7 @@ import PTree from './PTree.vue';
 import PI from '@/foundation/icons/PI.vue'
 import PIconButton from '@/inputs/buttons/icon-button/PIconButton.vue'
 import PButton from '@/inputs/buttons/button/PButton.vue'
-import PTextInput from '@/inputs/input/PTextInput.vue';
+import PTextInput from '@/inputs/input/text-input/PTextInput.vue';
 import PRadio from '@/inputs/radio/PRadio.vue'
 import { reactive, toRefs, computed } from 'vue';
 import {getTreeList, getTreeData, getTreeNode, getRecursiveInfo, getPermissionInfo} from './mock';
