+++ src/inputs/forms/field-group/PFieldGroup.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
 import PFieldGroup from './PFieldGroup.vue';
-import PTextInput from '@/inputs/input/PTextInput';
+import PTextInput from '@/inputs/input/text-input/PTextInput.vue';
 import { reactive, toRefs } from 'vue';
