+++ src/inputs/input/text-input/PTextInput.stories.mdx
@@ -1,18 +1,18 @@
 import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
 import Fuse from 'fuse.js'
-import PTextInput from '@/inputs/input/PTextInput.vue';
-import PButton from "@/inputs/buttons/button/PButton";
-import PTextEditor from "@/inputs/text-editor/PTextEditor";
-import PFieldGroup from "@/inputs/forms/field-group/PFieldGroup";
+import PTextInput from '@/inputs/input/text-input/PTextInput.vue';
+import PButton from "@/inputs/buttons/button/PButton.vue";
+import PTextEditor from "@/inputs/text-editor/PTextEditor.vue";
+import PFieldGroup from "@/inputs/forms/field-group/PFieldGroup.vue";
 import { reactive, toRefs } from 'vue';
 import {
     getTextInputArgTypes
-} from "@/inputs/input/story-helper";
-import {getTextInputMenu, getTextInputMenuWithMultiTypes} from "@/inputs/input/mock";
-import {INPUT_SIZE} from "@/inputs/input/type";
+} from "@/inputs/input/text-input/story-helper";
+import {getTextInputMenu, getTextInputMenuWithMultiTypes} from "@/inputs/input/text-input/mock";
+import {INPUT_SIZE} from "@/inputs/input/text-input/type";
 
 
-<Meta title='Inputs/Input' parameters={{
+<Meta title='Inputs/Input/Text Input' parameters={{
     design: {
         type: 'figma',
         url: 'https://www.figma.com/file/wq4wSowBcADBuUrMEZLz6i/SpaceONE-Console-Design?node-id=5191%3A2'
