+++ src/inputs/input/text-input/composables/use-selected-validation.ts
@@ -1,7 +1,7 @@
 import type { Ref } from 'vue';
 import { computed } from 'vue';
 
-import type { InputItem } from '@/inputs/input/type';
+import type { InputItem } from '@/inputs/input/text-input/type';
 
 interface UseSelectedValidationOptions {
     selected: Ref<InputItem[]>
