+++ src/others/console/modals/double-check-modal/PDoubleCheckModal.vue
@@ -46,7 +46,7 @@ import { reactive, toRefs } from 'vue';
 import PButtonModal from '@/feedbacks/modals/button-modal/PButtonModal.vue';
 import { SizeMapping } from '@/feedbacks/modals/type';
 import PFieldGroup from '@/inputs/forms/field-group/PFieldGroup.vue';
-import PTextInput from '@/inputs/input/PTextInput.vue';
+import PTextInput from '@/inputs/input/text-input/PTextInput.vue';
 import { makeProxy } from '@/util/composition-helpers';
